requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages


pandas
numpy
praw
configparser
apache-airflow-providers-amazon==8.17.0